move resuming sleeping runs onto the scheduler
